+++ example/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:code_scan/code_scan.dart';
-import 'package:google_mlkit_barcode_scanning/google_mlkit_barcode_scanning.dart';
 
 void main() async {
   runApp(const MyApp());
@@ -24,11 +23,11 @@ class _MyAppState extends State<MyApp> with TickerProviderStateMixin {
           CodeScanner(
             onScan: (code, details, controller) => setState(() => this.code = code),
             onScanAll: (codes, controller) => print('Codes: ' + codes.map((code) => code.rawValue).toString()),
-            formats: [ BarcodeFormat.qrCode ],
+            formats: const [ BarcodeFormat.qrCode ],
             once: true,
           ) : 
           GestureDetector(
-            onTap: () => setState(() => this.code = null),
+            onTap: () => setState(() => code = null),
             child: Container(
               color: Colors.transparent,
               child: Center(child: Text(code!)),